Optimization patch II
1. Using l2 socket instead of l3 socket and adding Ether() with each packet.
When we use l3 socket, it just gives us edge that it won't close the socket
after sending each packet with send(), sr() or sr1() command. But for our
case this improvement is not even enough. [Send 50 packets within 20ms window]

When we use l3 socket, scapy internally add mac address of the source and
dest. If it doesn't know the physical address (found through the ARP packets
which are sent periodically to update the routing table), it has to find
that first. So we are providing Ether() with the packet and sending it
a l2 socket.

2. Even with the improvement with l2 socket, we cannot send 50 packets
within a 20ms window. It takes around 40~70ms. So to improve this, we are
now sending raw(packet). When we use l2 or l3 socket, scapy internally
has to convert the packet (scapy format) to raw format before sending.
That's why it takes too much time to send the scapy packets. So we are
just preprocessing the packets beforehand and then sending them with
a burst. This now takes 2~10 ms to send 50 packets, which is below the
ICMP rate limit.